# Optimize for updates and eviction