--- a/Cargo.toml
+++ b/Cargo.toml
@@ -10,7 +10,6 @@ readme = "README.md"
 
 [dependencies]
 hdrsample = "3.0"
-twox-hash = "1.0"
 log = "0.3"
 ordermap = "0.2.10"