--- a/examples/multithread_stat.rs
+++ b/examples/multithread_stat.rs
@@ -38,7 +38,7 @@ fn main() {
     let loop_counter = metrics.counter("loop_counter".into());
     let loop_time_us = metrics.stat("loop_time_us".into());
     for work_done_tx in vec![work_done_tx0, work_done_tx1] {
-        let mut loop_counter = loop_counter.clone();
+        let loop_counter = loop_counter.clone();
         let mut loop_time_us = loop_time_us.clone();
         thread::spawn(move || {
             let mut prior: u64 = 0;
@@ -77,6 +77,6 @@ fn reporter<D>(interval: Duration, done: D, reporter: tacho::Reporter) -> BoxFut
     periodic.select(done).map(|_| {}).map_err(|_| {}).boxed()
 }
 
-fn print_report<R: tacho::Report>(report: &R) {
+fn print_report(report: &tacho::Report) {
     info!("\n{}", tacho::prometheus::format(report));
 }