--- a/src/report.rs
+++ b/src/report.rs
@@ -1,141 +1,109 @@
-use super::{CounterMap, GaugeMap, StatMap};
-use std::sync::{Arc, Mutex, MutexGuard};
+use super::{Key, Registry, CounterMap, GaugeMap};
+use hdrsample::Histogram;
+use ordermap::OrderMap;
+use std::sync::{Arc, Mutex};
+use std::sync::atomic::Ordering;
 
-pub fn new(counters: Arc<Mutex<CounterMap>>,
-           gauges: Arc<Mutex<GaugeMap>>,
-           stats: Arc<Mutex<StatMap>>)
-           -> Reporter {
-    Reporter {
-        counters: counters,
-        gauges: gauges,
-        stats: stats,
-    }
+type ReportCounterMap = OrderMap<Key, usize>;
+type ReportGaugeMap = OrderMap<Key, usize>;
+type ReportStatMap = OrderMap<Key, Histogram<usize>>;
+
+pub fn new(registry: Arc<Mutex<Registry>>) -> Reporter {
+    Reporter(registry)
 }
 
 #[derive(Clone)]
-pub struct Reporter {
-    counters: Arc<Mutex<CounterMap>>,
-    gauges: Arc<Mutex<GaugeMap>>,
-    stats: Arc<Mutex<StatMap>>,
-}
+pub struct Reporter(Arc<Mutex<Registry>>);
 
 impl Reporter {
     /// Obtains a read-only view of a metrics report without clearing the underlying state.
-    pub fn peek(&self) -> ReportPeek {
-        let counters = self.counters
-            .lock()
-            .expect("failed to obtain read lock for counters");
-        let gauges = self.gauges
-            .lock()
-            .expect("failed to obtain read lock for gauges");
-        let stats = self.stats
-            .lock()
-            .expect("failed to obtain read lock for stats");
-        ReportPeek {
-            counters: counters,
-            gauges: gauges,
-            stats: stats,
-        }
-    }
-
-    /// Obtains a Report and clears the underlying gauges and stats.
-    ///
-    /// Counters are copied and not cleared because counters are absolute and increasing.
-    pub fn take(&mut self) -> ReportTake {
-        // Copy counters.
-        let counters: CounterMap = {
-            let orig = self.counters
-                .lock()
-                .expect("failed to obtain write lock for counters");
-            let mut snap = CounterMap::default();
-            for (k, v) in orig.iter() {
-                snap.insert(k.clone(), *v);
-            }
-            snap
-        };
-
-        // Reset gauges.
-        let gauges = {
-            let mut orig = self.gauges
-                .lock()
-                .expect("failed to obtain write lock for gauges");
-            let mut snap = GaugeMap::default();
-            for (k, v) in orig.drain(..) {
-                snap.insert(k, v);
-            }
-            snap
-        };
-
-        // Reset stats.
+    pub fn peek(&self) -> Report {
+        let registry = self.0.lock().unwrap();
+        let counters = snap_counters(&registry.counters);
+        let gauges = snap_gauges(&registry.gauges);
         let stats = {
-            let mut orig = self.stats
-                .lock()
-                .expect("failed to obtain write lock for stats");
-            let mut snap = StatMap::default();
-            for (k, v) in orig.drain(..) {
-                snap.insert(k, v);
+            let mut snap = ReportStatMap::with_capacity(registry.stats.len());
+            for (k, v) in &registry.stats {
+                let v = v.lock().unwrap();
+                snap.insert(k.clone(), v.clone());
             }
             snap
         };
+        Report {
+            counters,
+            gauges,
+            stats,
+        }
+    }
 
-        ReportTake {
-            counters: counters,
-            gauges: gauges,
-            stats: stats,
+    /// Obtains a Report and removes unused metrics.
+    pub fn take(&mut self) -> Report {
+        let mut registry = self.0.lock().unwrap();
+        let counters = {
+            let snap = snap_counters(&registry.counters);
+            registry.counters.retain(|_, v| Arc::weak_count(v) > 0);
+            snap
+        };
+        let gauges = {
+            let snap = snap_gauges(&registry.gauges);
+            registry.gauges.retain(|_, v| Arc::weak_count(v) > 0);
+            snap
+        };
+        let stats = {
+            let mut snap = ReportStatMap::with_capacity(registry.stats.len());
+            for (k, ptr) in registry.stats.iter_mut() {
+                let mut orig = ptr.lock().unwrap();
+                snap.insert(k.clone(), orig.clone());
+                orig.clear();
+            }
+            registry.stats.retain(|_, v| Arc::weak_count(v) > 0);
+            snap
+        };
+        Report {
+            counters,
+            gauges,
+            stats,
         }
     }
 }
 
-pub trait Report {
-    fn is_empty(&self) -> bool;
-    fn len(&self) -> usize;
-    fn counters(&self) -> &CounterMap;
-    fn gauges(&self) -> &GaugeMap;
-    fn stats(&self) -> &StatMap;
+fn snap_counters(counters: &CounterMap) -> ReportCounterMap {
+    let mut snap = ReportCounterMap::with_capacity(counters.len());
+    for (k, v) in &*counters {
+        let v = v.load(Ordering::Acquire);
+        snap.insert(k.clone(), v);
+    }
+    snap
 }
 
-pub struct ReportPeek<'a> {
-    counters: MutexGuard<'a, CounterMap>,
-    gauges: MutexGuard<'a, GaugeMap>,
-    stats: MutexGuard<'a, StatMap>,
-}
-impl<'a> Report for ReportPeek<'a> {
-    fn counters(&self) -> &CounterMap {
-        &self.counters
+fn snap_gauges(gauges: &GaugeMap) -> ReportGaugeMap {
+    let mut snap = ReportGaugeMap::with_capacity(gauges.len());
+    for (k, v) in &*gauges {
+        let v = v.load(Ordering::Acquire);
+        snap.insert(k.clone(), v);
     }
-    fn gauges(&self) -> &GaugeMap {
-        &self.gauges
-    }
-    fn stats(&self) -> &StatMap {
-        &self.stats
-    }
-    fn is_empty(&self) -> bool {
-        self.counters.is_empty() && self.gauges.is_empty() && self.stats.is_empty()
-    }
-    fn len(&self) -> usize {
-        self.counters.len() + self.gauges.len() + self.stats.len()
-    }
+    snap
 }
 
-pub struct ReportTake {
-    pub counters: CounterMap,
-    pub gauges: GaugeMap,
-    pub stats: StatMap,
+pub struct Report {
+    counters: ReportCounterMap,
+    gauges: ReportGaugeMap,
+    stats: ReportStatMap,
 }
-impl Report for ReportTake {
-    fn counters(&self) -> &CounterMap {
+impl Report {
+    pub fn counters(&self) -> &ReportCounterMap {
         &self.counters
     }
-    fn gauges(&self) -> &GaugeMap {
+    pub fn gauges(&self) -> &ReportGaugeMap {
         &self.gauges
     }
-    fn stats(&self) -> &StatMap {
+    pub fn stats(&self) -> &ReportStatMap {
         &self.stats
     }
-    fn is_empty(&self) -> bool {
+    pub fn is_empty(&self) -> bool {
         self.counters.is_empty() && self.gauges.is_empty() && self.stats.is_empty()
     }
-    fn len(&self) -> usize {
+    pub fn len(&self) -> usize {
         self.counters.len() + self.gauges.len() + self.stats.len()
     }
 }