Improve metric updates benchmarks subtantially by using AtomicUsize counters
and gauges. Furthermore, each stat is now protected by its own mutex so that
the entire stat registry need not be locked for a simple measurement.

However, the use of atomics prevents gauges from being evicted from the
registry after a report. In order to fix this, each Counter, Gauge, & Stat now
maintains a Weak reference back to the entry in the central metrics registry.
This allows the Reporter to evict all metrics that are no longer referenced by
a receiver (including counters and stats) without losing values before they are
reported.  The Report type has been simplified so that reports hold a stable
copy of data (which is now necessary that each metric is its own reference).

Other improvements include:
- removal of the twox hash, which negatively impacts performance of updates
- metric and label keys now must be `&'string str` (label values may still be dynamic)

The goal of these changes is to prioritize value updates over metric creation
over reporting (which is typically performed in an auxiliary thread).

Comparing previous behavior, we take some penalty in metrics creation over a
substantial improvement in updates:
```
:; cargo benchcmp ver-benchmarks ver-optimize-updates-and-gc
 name                               ver-benchmarks ns/iter  ver-optimize-updates-and-gc ns/iter  diff ns/iter   diff %   speedup
 tests::bench_counter_create        72                      91                                             19   26.39%    x 0.79
 tests::bench_counter_create_x1000  120,375                 157,364                                    36,989   30.73%    x 0.76
 tests::bench_counter_update        70                      20                                            -50  -71.43%    x 3.50
 tests::bench_counter_update_x1000  106,813                 619                                      -106,194  -99.42%  x 172.56
 tests::bench_gauge_create          68                      86                                             18   26.47%    x 0.79
 tests::bench_gauge_create_x1000    118,357                 157,145                                    38,788   32.77%    x 0.75
 tests::bench_gauge_update          62                      16                                            -46  -74.19%    x 3.88
 tests::bench_gauge_update_x1000    107,805                 613                                      -107,192  -99.43%  x 175.86
 tests::bench_stat_add_x1000        68,425                  2,581                                     -65,844  -96.23%   x 26.51
 tests::bench_stat_create           67                      84                                             17   25.37%    x 0.80
 tests::bench_stat_create_x1000     117,878                 152,086                                    34,208   29.02%    x 0.78
 tests::bench_stat_update           71                      41                                            -30  -42.25%    x 1.73
 tests::bench_stat_update_x1000     129,666                 2,981                                    -126,685  -97.70%   x 43.50
```

```
counter "target/release/examples/multithread_${kind}"  0.75s user 0.00s system 195% cpu 0.389 total
gauge   "target/release/examples/multithread_${kind}"  0.74s user 0.00s system 196% cpu 0.380 total
stat    "target/release/examples/multithread_${kind}"  1.73s user 8.11s system 105% cpu 9.332 total
```
(down from 4-7s)

Based on https://github.com/linkerd/tacho/pull/17.  This will be rebased from master once that branch is merged.