@olix0r 2017-05-28
With twox vs without:
```
:; cargo benchcmp ver-optimize-updates-and-gc{-twox,}   
 name                               ver-optimize-updates-and-gc-twox ns/iter  ver-optimize-updates-and-gc ns/iter  diff ns/iter   diff %  speedup 
 tests::bench_counter_create        101                                       91                                            -10   -9.90%   x 1.11 
 tests::bench_counter_create_x1000  173,155                                   157,364                                   -15,791   -9.12%   x 1.10 
 tests::bench_counter_update        20                                        20                                              0    0.00%   x 1.00 
 tests::bench_counter_update_x1000  611                                       619                                             8    1.31%   x 0.99 
 tests::bench_gauge_create          98                                        86                                            -12  -12.24%   x 1.14 
 tests::bench_gauge_create_x1000    179,785                                   157,145                                   -22,640  -12.59%   x 1.14 
 tests::bench_gauge_update          15                                        16                                              1    6.67%   x 0.94 
 tests::bench_gauge_update_x1000    650                                       613                                           -37   -5.69%   x 1.06 
 tests::bench_stat_add_x1000        2,687                                     2,581                                        -106   -3.94%   x 1.04 
 tests::bench_stat_create           101                                       84                                            -17  -16.83%   x 1.20 
 tests::bench_stat_create_x1000     173,609                                   152,086                                   -21,523  -12.40%   x 1.14 
 tests::bench_stat_update           40                                        41                                              1    2.50%   x 0.98 
 tests::bench_stat_update_x1000     3,101                                     2,981                                        -120   -3.87%   x 1.04 
```