@stevej 2017-05-28
For the benchmark to be valid, it needs to ensure the key lengths are what we expect to see in production since hash functions are often very sensitive to key length; Do we plan to see our keys grow in size to be close to the size we get from a finagle process? I did benchmark xxHash with 32-byte keys and should have added that to the benchmarks in this suite. :-(