@olix0r 2017-05-29
@stevej in our case, the key of the hash is an object and not a flat string -- i'll change the tests to use a somewhat longer stat name